fixes #30 Added half-height cell types. This required introducing another property, translucency, to the cell types.
* Made the Player able to place half-sized conterparts of existing cell types.
Of course it costs a half material of full height one, so we need to store
bricks as fixed point number in binary.
* Made the Player capable of climbing steps lower than 0.5 meters.
This way you do not need to jump to step over half-height cells.
* Reduced Player's jump power because you can climb up 1-unit cell height with
lower jump height than before.
* Added namespace braces to World.cpp because it bothers me to wite redundant
dxtest:: prior to every symbol.
* Moved some of World's methods from World.h to World.cpp, because they were too
complex to be inline expanded.